Move Table definitions to mozak_stark
