+++ circuits/src/stark/mozak_stark.rs
@@ -1,10 +1,77 @@
 use plonky2::field::extension::Extendable;
+use plonky2::field::types::Field;
 use plonky2::hash::hash_types::RichField;
 
 use crate::cpu::stark::CpuStark;
+use crate::cross_table_lookup::{Column, CrossTableLookup};
+use crate::{cpu, rangecheck};
 #[derive(Clone, Default)]
 pub struct MozakStark<F: RichField + Extendable<D>, const D: usize> {
     pub cpu_stark: CpuStark<F, D>,
 }
 
 pub(crate) const NUM_TABLES: usize = 1;
+
+#[derive(Debug, Copy, Clone)]
+pub enum TableKind {
+    Cpu = 0,
+    RangeCheck = 1,
+}
+
+#[derive(Clone, Debug)]
+#[allow(unused)]
+pub struct Table<F: Field> {
+    pub(crate) kind: TableKind,
+    pub(crate) columns: Vec<Column<F>>,
+    pub(crate) filter_column: Option<Column<F>>,
+}
+
+impl<F: Field> Table<F> {
+    pub fn new(kind: TableKind, columns: Vec<Column<F>>, filter_column: Option<Column<F>>) -> Self {
+        Self {
+            kind,
+            columns,
+            filter_column,
+        }
+    }
+}
+
+/// Represents a range check trace table in the Mozak VM.
+pub struct RangeCheckTable<F: Field>(Table<F>);
+
+/// Represents a cpu trace table in the Mozak VM.
+pub struct CpuTable<F: Field>(Table<F>);
+
+impl<F: Field> RangeCheckTable<F> {
+    #[allow(clippy::new_ret_no_self)]
+    pub fn new(columns: Vec<Column<F>>, filter_column: Option<Column<F>>) -> Table<F> {
+        Table::new(TableKind::RangeCheck, columns, filter_column)
+    }
+}
+
+impl<F: Field> CpuTable<F> {
+    #[allow(clippy::new_ret_no_self)]
+    pub fn new(columns: Vec<Column<F>>, filter_column: Option<Column<F>>) -> Table<F> {
+        Table::new(TableKind::Cpu, columns, filter_column)
+    }
+}
+pub trait Lookups<F: Field> {
+    fn lookups() -> CrossTableLookup<F>;
+}
+
+pub struct RangecheckCpuTable<F: Field>(CrossTableLookup<F>);
+
+impl<F: Field> Lookups<F> for RangecheckCpuTable<F> {
+    fn lookups() -> CrossTableLookup<F> {
+        CrossTableLookup::new(
+            vec![CpuTable::new(
+                cpu::columns::data_for_rangecheck(),
+                Some(cpu::columns::filter_for_rangecheck()),
+            )],
+            RangeCheckTable::new(
+                rangecheck::columns::data_for_cpu(),
+                Some(rangecheck::columns::filter_for_cpu()),
+            ),
+        )
+    }
+}
